float parset

+++ app/src/main/java/com/example/damian/kinematicscalculatorvs3/adapters/AdapterForwardValue.java
@@ -126,10 +126,10 @@ public class AdapterForwardValue extends ArrayAdapter<ModelKinematicsForwardValu
 
                         ModelKinematicsForwardValueJoin model = new ModelKinematicsForwardValueJoin(position);
                         model.setTv_lp(modelKinematicsForwardValueJoin.getTv_lp());
-                        model.setEt_alpha(Integer.parseInt(listViewHolder.editalpha.getText().toString()));
-                        model.setEt_a(Integer.parseInt(listViewHolder.edita.getText().toString()));
-                        model.setEt_theta(Integer.parseInt(listViewHolder.edittheta.getText().toString()));
-                        model.setEt_d(Integer.parseInt(listViewHolder.editd.getText().toString()));
+                        model.setEt_alpha(Float.parseFloat(listViewHolder.editalpha.getText().toString()));
+                        model.setEt_a(Float.parseFloat(listViewHolder.edita.getText().toString()));
+                        model.setEt_theta(Float.parseFloat(listViewHolder.edittheta.getText().toString()));
+                        model.setEt_d(Float.parseFloat(listViewHolder.editd.getText().toString()));
 
                         StaticVolumesKinematicsForwardValue.setOneModel(model);
                         return false;
@@ -160,9 +160,9 @@ public class AdapterForwardValue extends ArrayAdapter<ModelKinematicsForwardValu
                     public boolean onEditorAction(TextView v, int actionId, KeyEvent event) {
 
                         ModelKinematicsForwardValueEffector model = new ModelKinematicsForwardValueEffector(position);
-                        model.setEt_x(Integer.parseInt(listViewHolder.editX.getText().toString()));
-                        model.setEt_y(Integer.parseInt(listViewHolder.editY.getText().toString()));
-                        model.setEt_z(Integer.parseInt(listViewHolder.editZ.getText().toString()));
+                        model.setEt_x(Float.parseFloat(listViewHolder.editX.getText().toString()));
+                        model.setEt_y(Float.parseFloat(listViewHolder.editY.getText().toString()));
+                        model.setEt_z(Float.parseFloat(listViewHolder.editZ.getText().toString()));
 
                         StaticVolumesKinematicsForwardValue.setOneModel(model);
                         return false;
